MINOR Add modularization to ajax code
All of the AJAX code will be modularized. As of now, just the menu was
ported.

The menu will be added from a list of interest. It's interesting but not
sure how necessary it will be to have a menu based on a template. Might
have bugs.

The files changed their names thanks to the new architecture that uses
templates. The first implementation was the menu.