MINOR Reorganized code structure
The code was reorganized so any individual with less skill than myself -
which sais a lot - can edit if he desires so.

There are actions to be taken and libriries it depends on.

Every idea hides behind libraries. Every interaction hides behind
actions that are taken. It might be blurry or you might disagree upon
seeing the code but that is what I take is correct.

I also removed some legacy code and files. Some stuff was renamed and
rearranged.