MINOR Change to variables design
The space for each individual variable now increased 1 fr. The overal
grid of the ranger and bubbles increased one as consequence.

Each variable now has the color of the element - more or less. It also
has a monospaced font and centralization.

A css file was created to hold these changes.