MINOR Change to graph positioning and content
The graph shows now only concentration of external potassium versus the
potential energy of potassion in human membrane.

This change is the beginning of the entire thing as the chosen ion will
be shown only. And the graph will reflect that ion values in control
panel for the variables.

It's position changed as well. More changes to come.