MINOR Change to pre-teste page
The page is now in template form. AJAX calls a PHP template and
configure each question-answer.

A new action for this was required. This action AJAX the content in a
non-destructive way.

The design also changed a little.

TODO: might be interesting to change the grid because mobile doesn't
work that well. Might be unusable. Despite it, it's supposed to be run
on desktop.