MINOR Change to tests design
The tests suffered a reformulation to be viewed more as a block (code).
The visual was also changed.

Reformulation of the CSS was necessary. So it was an oportunity to
change it while redesigning the UI of the tests.

The questions and alternatives were finally added properly for both pre
and pos-tests. The correct answer is the first - always.

There is a need to implement a system do save and verify each question
at the right time. Then, it will be finalized - at the very basic.

Corrected bugs:
- AJAX with GET had a problem with strings containing characters used in
  URI. So the text is encoded in JavaScript and decoded in PHP. It's now
  fixed.
- The radio buttons increased in size with multiple lines of text being
  present. With CSS "height: fit-content", those problems were fixed.