MINOR Change in graph UI and globals addition
The graph now changes in accordance to the ion being fiddled with.

The graph now supports zoom and dislocation of the curve.

Some changes in names were made of the variables. There is also support
for globals now. In the future, everything universally accessed by
JavaScript, will be in the form "_NAME" in the lib_gloals.js file.