MINOR Add support for tables design
A generic design for tables was made. It consider a tr with th as
headers.It also considers the first column to be a different header -
but made with td as the rest of the row will be normal cross data.

The only table so far was made to be compatible.