MINOR Convert template PHP 2 JS and make all array
The code used as template from PHP was functional. The problem: the fact
it required ajax meant the PHP could run concurrent to javascript code
and strange stuff could happen. It broke the 'sequential' caracteristic.

The solution was to make the template into a javascript function. So,
that means it is faster, doens't require PHP and runs without bugs and
doesn't need tricks to slow down certain parts of the code.

This PHP 2 JS thing was made for tests only. TODO for menu entries.

Objects are also non-guaranteed to be in order, only arrays. So,
everything for tests converted to Array. Basically, from {} to []. And
it affected both menu and tests that used AJAX.

This change was meant to help with the assynchronous and concurrent
running of code. Not exactly necessary but makes things faster.

Some division in globals was made for globals and helpers of arrays.
More is needed to be done in the future.

Some other changes were made to make the code work and compliant with
the project. Like some renaming, simplifications, new files...