MINOR Add low level library
This library is reserved for low code implementations. Meaning stuff
necessary that are considered at the bottom of a system or language.

It currently only implements the sleep function. The time is
configurable at the time you invoke the function (miliseconds).

This sleep function bases itself on the premisse that JavaScript is
monothreaded (it is) and will always be.