PATCH Change interval of values in experiment
The values - range and steps - for the ions were changed per request to
better inform the desired result in normal situation within the
phisiological aspect of the human body. It also helps the students to
observe the differences in changes - being small or big - of
eletrochemical potential and membrane potential thanks to changes in
permeability, concentration and condutance.

There are also small changes to text and font sizes.